requirements.txt
# --- Web Framework & Auth ---
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Caching==2.3.1
Flask-Cors==6.0.0
Werkzeug==3.1.3
python-dotenv==1.0.1

# --- Geoprocessamento Core ---
numpy>=1.26.0
pandas==2.2.3
geopandas==1.1.0
rasterio==1.4.3
shapely==2.1.1
fiona==1.10.1
pyproj==3.7.1
geojson==3.2.0

# --- Satélite & STAC ---
sentinelhub==3.11.1
pystac==1.13.0
pystac-client==0.8.6
rio-tiler==7.8.0

# --- Imagem & Machine Learning ---
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
pillow==11.2.1
tifffile==2025.5.10

# --- Performance & Armazenamento ---
pyarrow
fastparquet

# --- Utilidades ---
requests==2.32.3
urllib3==2.4.0
tqdm==4.67.1